add new test for delete file
